+++ styles/font.ts
@@ -1,13 +1,12 @@
-export const FONT_SIZE= {
+export const FONT_SIZE = {
     title: "144px",
     subTitle: "56px",
     xxl: "48px",
-	xl: "32px", 
-	lg: "26px", 
-	bg: "24px", 
-	md: "20px", 
-	sm: "18px", 
-	xs: "16px", 
-	xxs: "15px", 
-}as const;
-
+    xl: "32px",
+    lg: "26px",
+    bg: "24px",
+    md: "20px",
+    sm: "18px",
+    xs: "16px",
+    xxs: "15px",
+} as const
